+++ packages/scenario/src/client/init.ts
@@ -1,50 +1,18 @@
 import type { JRResourceService } from '@getodk/common/jr-resources/JRResourceService.ts';
-import type { XFormsElement } from '@getodk/common/test/fixtures/xform-dsl/XFormsElement.ts';
 import type {
-	EngineConfig,
 	FormResource,
+	InstanceAttachmentsConfig,
+	LoadFormOptions,
+	LoadFormSuccessResult,
+	LoadFormWarningResult,
+	MissingResourceBehavior,
 	OpaqueReactiveObjectFactory,
 	RootNode,
 } from '@getodk/xforms-engine';
-import { initializeForm } from '@getodk/xforms-engine';
+import { createInstance } from '@getodk/xforms-engine';
 import type { Owner } from 'solid-js';
-import { createRoot, getOwner, runWithOwner } from 'solid-js';
-import type { MissingResourceBehavior } from '../../../xforms-engine/dist/client/constants';
-import { FormDefinitionResource } from '../jr/resource/FormDefinitionResource.ts';
-
-/**
- * @todo It's anticipated that this will be expanded to support the various ways
- * that JavaRosa's tests load form resources. This initially supports the (also
- * ported) in-source DSL. We'll likely also need to support loading fixtures by
- * file name (and various ambiguities that come with that, which we may want to
- * disambiguate as we port).
- */
-export type TestFormResource = FormDefinitionResource | XFormsElement;
-
-const isPathResource = (resource: TestFormResource): resource is FormDefinitionResource => {
-	return resource instanceof FormDefinitionResource;
-};
-
-const isXFormsElement = (resource: TestFormResource): resource is XFormsElement => {
-	return typeof (resource as XFormsElement).asXml === 'function';
-};
-
-export const getFormResource = async (
-	testFormResource: TestFormResource
-	// It's also anticipated that we will need to perform async IO as we expand
-	// the `TestFormResource` type.
-	// eslint-disable-next-line @typescript-eslint/require-await
-): Promise<FormResource> => {
-	if (isPathResource(testFormResource)) {
-		return testFormResource.textContents;
-	}
-
-	if (isXFormsElement(testFormResource)) {
-		return testFormResource.asXml();
-	}
-
-	throw new Error('Unknown test form resource');
-};
+import { createRoot } from 'solid-js';
+import { getAssertedOwner, runInSolidScope } from './solid-helpers.ts';
 
 /**
  * @todo Currently we stub resource fetching. We can address this as needed
@@ -54,46 +22,52 @@ const fetchFormDefinitionStub: typeof fetch = () => {
 	throw new Error('TODO: fetching form definition not implemented');
 };
 
-export interface InitializeTestFormOptions {
+export interface TestFormOptions {
 	readonly resourceService: JRResourceService;
 	readonly missingResourceBehavior: MissingResourceBehavior;
 	readonly stateFactory: OpaqueReactiveObjectFactory;
+	readonly instanceAttachments: InstanceAttachmentsConfig;
 }
 
 const defaultConfig = {
 	fetchFormDefinition: fetchFormDefinitionStub,
-} as const satisfies Omit<EngineConfig, 'stateFactory'>;
+} as const satisfies LoadFormOptions;
+
+// prettier-ignore
+export type InitializableForm =
+	| LoadFormSuccessResult
+	| LoadFormWarningResult;
 
 interface InitializedTestForm {
+	readonly form: InitializableForm;
 	readonly instanceRoot: RootNode;
 	readonly owner: Owner;
 	readonly dispose: VoidFunction;
 }
 
 export const initializeTestForm = async (
-	testForm: TestFormResource,
-	options: InitializeTestFormOptions
+	formResource: FormResource,
+	options: TestFormOptions
 ): Promise<InitializedTestForm> => {
 	return createRoot(async (dispose) => {
-		const owner = getOwner();
+		const owner = getAssertedOwner();
 
-		if (owner == null) {
-			throw new Error('Must have reactive context owner');
-		}
-
-		const formResource = await getFormResource(testForm);
-		const instanceRoot = await runWithOwner(owner, async () => {
-			return initializeForm(formResource, {
-				config: {
+		const { formResult: form, root: instanceRoot } = await runInSolidScope(owner, async () => {
+			return createInstance(formResource, {
+				form: {
 					...defaultConfig,
 					fetchFormAttachment: options.resourceService.handleRequest,
 					missingResourceBehavior: options.missingResourceBehavior,
+				},
+				instance: {
 					stateFactory: options.stateFactory,
+					instanceAttachments: options.instanceAttachments,
 				},
 			});
-		})!;
+		});
 
 		return {
+			form,
 			instanceRoot,
 			owner,
 			dispose,