+++ packages/scenario/vitest.config.ts
@@ -55,8 +55,8 @@ export default defineConfig(({ mode }) => {
 			esbuildOptions: {
 				target: 'esnext',
 			},
-			exclude: ['@getodk/xforms-engine'],
 			force: true,
+			include: ['papaparse'],
 		},
 		resolve: {
 			alias: {
@@ -67,7 +67,7 @@ export default defineConfig(({ mode }) => {
 		test: {
 			browser: {
 				enabled: BROWSER_ENABLED,
-				name: BROWSER_NAME!,
+				instances: BROWSER_NAME != null ? [{ browser: BROWSER_NAME }] : [],
 				provider: 'playwright',
 				headless: true,
 				screenshotFailures: false,
@@ -79,10 +79,9 @@ export default defineConfig(({ mode }) => {
 			deps: {
 				optimizer: {
 					web: {
-						// Prevent loading multiple instances of Solid. This deviates from
-						// most of the recommendations provided by Solid and related tooling,
-						// as Vitest's interfaces have since changed. But it does seem to be
-						// the appropriate solution (at least for our usage).
+						// Prevent loading multiple instances of Solid.
+						//
+						// (Copypasta from `@getodk/common`)
 						exclude: ['solid-js'],
 					},
 				},
@@ -93,6 +92,20 @@ export default defineConfig(({ mode }) => {
 			setupFiles: ['./src/vitest/setup.ts'],
 
 			reporters: process.env.GITHUB_ACTIONS ? ['default', 'github-actions'] : 'default',
+
+			server: {
+				deps: {
+					/**
+					 * Added to resolve a `TypeError: Cannot read properties of undefined (reading 'registerGraph')`
+					 * error in `solid-js/store` during tests, which occurred because SolidJS dev tools (`DEV$1`) were not
+					 * properly initialized in the `jsdom` environment when dependencies were pre-bundled.
+					 *
+					 * It maintains test behavior closer to a real browser runtime, avoiding pre-bundling quirks. It might
+					 * increase test startup time slightly due to skipping pre-bundling optimizations.
+					 */
+					inline: TEST_ENVIRONMENT === 'jsdom' ? ['solid-js'] : [],
+				},
+			},
 		},
 	};
 });